+++ src/moonblade/agg/mod.rs
@@ -4,6 +4,6 @@ mod stats;
 mod window;
 
 pub use aggregators::{CovarianceWelford, Welford};
-pub use program::{AggregationProgram, GroupAggregationProgram};
+pub use program::{AggregationProgram, GroupAggregationProgram, GroupPivotAggregationProgram};
 pub use stats::Stats;
 pub use window::WindowAggregationProgram;